+++ 1.20/Common/src/main/java/fuzs/mutantmonsters/client/gui/screens/CreeperMinionTrackerScreen.java
@@ -5,6 +5,7 @@ import fuzs.mutantmonsters.MutantMonsters;
 import fuzs.mutantmonsters.network.client.C2SCreeperMinionNameMessage;
 import fuzs.mutantmonsters.network.client.C2SCreeperMinionTrackerMessage;
 import fuzs.mutantmonsters.world.entity.CreeperMinion;
+import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.AbstractWidget;
@@ -13,10 +14,23 @@ import net.minecraft.client.gui.components.EditBox;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.CommonComponents;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.Mth;
+
+import java.text.DecimalFormat;
+import java.text.DecimalFormatSymbols;
+import java.util.Locale;
 
 public class CreeperMinionTrackerScreen extends Screen {
     private static final ResourceLocation TEXTURE_LOCATION = MutantMonsters.id("textures/gui/creeper_minion_tracker.png");
+    private static final MutableComponent HEALTH_COMPONENT = Component.translatable("gui.mutantmonsters.creeper_minion_tracker.health");
+    private static final MutableComponent EXPLOSION_COMPONENT = Component.translatable("gui.mutantmonsters.creeper_minion_tracker.explosion");
+    private static final MutableComponent BLAST_RADIUS_COMPONENT = Component.translatable("gui.mutantmonsters.creeper_minion_tracker.blast_radius");
+    private static final DecimalFormat DECIMAL_FORMAT = Util.make(new DecimalFormat("#.0"), decimalFormat -> {
+        decimalFormat.setDecimalFormatSymbols(DecimalFormatSymbols.getInstance(Locale.ROOT));
+    });
+
     private final int imageWidth = 176;
     private final int imageHeight = 166;
     private final CreeperMinion creeperMinion;
@@ -57,17 +71,17 @@ public class CreeperMinionTrackerScreen extends Screen {
             this.canDestroyBlocks = !this.canDestroyBlocks;
             MutantMonsters.NETWORK.sendToServer(new C2SCreeperMinionTrackerMessage(this.creeperMinion, 0, this.canDestroyBlocks));
             button.setMessage(this.canDestroyBlocks());
-        }).bounds(this.leftPos + 8, this.topPos + this.imageHeight - 78, buttonWidth * 2 + 4, 20).build());
+        }).bounds(this.leftPos + 8, this.topPos + this.imageHeight - 75, buttonWidth * 2 + 4, 20).build());
         this.addRenderableWidget(Button.builder(this.alwaysShowName(), (button) -> {
             this.alwaysShowName = !this.alwaysShowName;
             MutantMonsters.NETWORK.sendToServer(new C2SCreeperMinionTrackerMessage(this.creeperMinion, 1, this.alwaysShowName));
             button.setMessage(this.alwaysShowName());
-        }).bounds(this.leftPos + 8, this.topPos + this.imageHeight - 54, buttonWidth * 2 + 4, 20).build());
+        }).bounds(this.leftPos + 8, this.topPos + this.imageHeight - 51, buttonWidth * 2 + 4, 20).build());
         this.addRenderableWidget(Button.builder(this.canRideOnShoulder(), (button) -> {
             this.canRideOnShoulder = !this.canRideOnShoulder;
             MutantMonsters.NETWORK.sendToServer(new C2SCreeperMinionTrackerMessage(this.creeperMinion, 2, this.canRideOnShoulder));
             button.setMessage(this.canRideOnShoulder());
-        }).bounds(this.leftPos + 8, this.topPos + this.imageHeight - 30, buttonWidth * 2 + 4, 20).build());
+        }).bounds(this.leftPos + 8, this.topPos + this.imageHeight - 27, buttonWidth * 2 + 4, 20).build());
         if (!this.creeperMinion.isOwnedBy(this.minecraft.player)) {
             this.renderables.stream().filter(AbstractWidget.class::isInstance).map(AbstractWidget.class::cast).forEach(widget -> widget.active = false);
         }
@@ -124,20 +138,17 @@ public class CreeperMinionTrackerScreen extends Screen {
         this.renderBackground(guiGraphics);
         RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
         guiGraphics.blit(TEXTURE_LOCATION, this.leftPos, this.topPos, 0, 0, this.imageWidth, this.imageHeight);
-        int health = (int) (this.creeperMinion.getHealth() * 150.0F / this.creeperMinion.getMaxHealth());
-        guiGraphics.blit(TEXTURE_LOCATION, this.leftPos + 13, this.topPos + 16, 0, this.imageHeight, health, 6);
+        guiGraphics.blit(TEXTURE_LOCATION, this.leftPos + 15, this.topPos + 16, 0, 166, 146, 5);
+        float healthProgress = Mth.clamp(this.creeperMinion.getHealth() / this.creeperMinion.getMaxHealth(), 0.0F, 1.0F);
+        guiGraphics.blit(TEXTURE_LOCATION, this.leftPos + 15, this.topPos + 16, 0, 171, (int) (healthProgress * 146.0F), 5);
         guiGraphics.drawString(this.font, this.title, this.leftPos + this.titleLabelX, this.topPos + this.titleLabelY, 4210752, false);
         this.name.render(guiGraphics, mouseX, mouseY, partialTicks);
-        guiGraphics.drawString(this.font, Component.translatable("gui.mutantmonsters.creeper_minion_tracker.health"), (this.leftPos + 13), (this.topPos + 28), 4210752, false);
-        guiGraphics.drawString(this.font, Component.translatable("gui.mutantmonsters.creeper_minion_tracker.explosion"), (this.leftPos + 13), (this.topPos + 48), 4210752, false);
-        guiGraphics.drawString(this.font, Component.translatable("gui.mutantmonsters.creeper_minion_tracker.blast_radius"), (this.leftPos + 13), (this.topPos + 68), 4210752, false);
-        StringBuilder builder = new StringBuilder();
-        builder.append(this.creeperMinion.getHealth() / 2.0F).append(" / ").append(this.creeperMinion.getMaxHealth() / 2.0F);
-        guiGraphics.drawCenteredString(this.font, builder.toString(), this.leftPos + this.imageWidth / 2 + 38, this.topPos + 30, 16777215);
-        guiGraphics.drawCenteredString(this.font, this.creeperMinion.canExplodeContinuously() ? Component.translatable("gui.mutantmonsters.creeper_minion_tracker.explosion.continuous") : Component.translatable("gui.mutantmonsters.creeper_minion_tracker.explosion.one_time"), this.leftPos + this.imageWidth / 2 + 38, this.topPos + 50, 16777215);
-        int temp = (int) (this.creeperMinion.getExplosionRadius() * 10.0F);
-        builder = new StringBuilder().append(temp / 10.0F);
-        guiGraphics.drawCenteredString(this.font, builder.toString(), this.leftPos + this.imageWidth / 2 + 38, this.topPos + 70, 16777215);
+        guiGraphics.drawString(this.font, HEALTH_COMPONENT, (this.leftPos + 13), (this.topPos + 31), 4210752, false);
+        guiGraphics.drawString(this.font, EXPLOSION_COMPONENT, (this.leftPos + 13), (this.topPos + 51), 4210752, false);
+        guiGraphics.drawString(this.font, BLAST_RADIUS_COMPONENT, (this.leftPos + 13), (this.topPos + 71), 4210752, false);
+        guiGraphics.drawCenteredString(this.font, String.format("%s / %s", DECIMAL_FORMAT.format(this.creeperMinion.getHealth()), DECIMAL_FORMAT.format(this.creeperMinion.getMaxHealth())), this.leftPos + this.imageWidth / 2 + 38, this.topPos + 31, 16777215);
+        guiGraphics.drawCenteredString(this.font, this.creeperMinion.canExplodeContinuously() ? Component.translatable("gui.mutantmonsters.creeper_minion_tracker.explosion.continuous") : Component.translatable("gui.mutantmonsters.creeper_minion_tracker.explosion.one_time"), this.leftPos + this.imageWidth / 2 + 38, this.topPos + 51, 16777215);
+        guiGraphics.drawCenteredString(this.font, DECIMAL_FORMAT.format(this.creeperMinion.getExplosionRadius()), this.leftPos + this.imageWidth / 2 + 38, this.topPos + 71, 16777215);
         super.render(guiGraphics, mouseX, mouseY, partialTicks);
     }