+++ 1.20/Common/src/main/java/fuzs/mutantmonsters/client/renderer/EndersoulHandRenderer.java
@@ -8,6 +8,7 @@ import fuzs.mutantmonsters.client.model.EndersoulHandModel;
 import fuzs.puzzleslib.api.client.init.v1.DynamicBuiltinItemRenderer;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.ItemRenderer;
 import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.client.resources.model.ModelResourceLocation;
@@ -25,12 +26,14 @@ public class EndersoulHandRenderer implements DynamicBuiltinItemRenderer {
     private EndersoulHandModel enderSoulHandModel;
 
     @Override
-    public void renderByItem(ItemStack stack, ItemDisplayContext mode, PoseStack matrices, MultiBufferSource vertexConsumers, int light, int overlay) {
+    public void renderByItem(ItemStack stack, ItemDisplayContext mode, PoseStack matrices, MultiBufferSource multiBufferSource, int light, int overlay) {
         matrices.pushPose();
         float ageInTicks = (float) this.minecraft.player.tickCount + this.minecraft.getFrameTime();
         this.enderSoulHandModel.setAngles();
-        VertexConsumer ivertexbuilder = ItemRenderer.getFoilBufferDirect(vertexConsumers, MutantRenderTypes.energySwirl(ENDERSOUL_HAND_TEXTURE, ageInTicks * 0.008F, ageInTicks * 0.008F), true, stack.hasFoil());
-        this.enderSoulHandModel.renderToBuffer(matrices, ivertexbuilder, 15728880, OverlayTexture.NO_OVERLAY, 0.9F, 0.3F, 1.0F, 1.0F);
+        RenderType renderType = MutantRenderTypes.energySwirl(ENDERSOUL_HAND_TEXTURE, ageInTicks * 0.008F, ageInTicks * 0.008F);
+        // ignore enchanting glint, it looks bad
+        VertexConsumer vertexConsumer = multiBufferSource.getBuffer(renderType);
+        this.enderSoulHandModel.renderToBuffer(matrices, vertexConsumer, 15728880, OverlayTexture.NO_OVERLAY, 0.9F, 0.3F, 1.0F, 1.0F);
         matrices.popPose();
     }