+++ 1.20/Common/src/main/java/fuzs/mutantmonsters/client/renderer/HulkHammerModels.java
@@ -1,0 +1,9 @@
+package fuzs.mutantmonsters.client.renderer;
+
+import fuzs.mutantmonsters.MutantMonsters;
+import net.minecraft.client.resources.model.ModelResourceLocation;
+
+public class HulkHammerModels {
+    public static final ModelResourceLocation HULK_HAMMER_IN_HAND_MODEL = new ModelResourceLocation(MutantMonsters.id("hulk_hammer_in_hand"), "inventory");
+    public static final ModelResourceLocation HULK_HAMMER_ITEM_MODEL = new ModelResourceLocation(MutantMonsters.id("hulk_hammer"), "inventory");
+}