+++ 1.20/CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v8.0.1-1.20.1] - 2023-07-01
+### Added
+- Added beautiful new textures in Minecraft's new art style provided by [tdstress](https://www.curseforge.com/members/tdstress)
+### Changed
+- Some adjustment to the creeper minion interaction screen
+### Fixed
+- Fixed a rare issue where rendering a teleporting mutant enderman could crash the game
+- The ender soul hand held model no longer shows an enchantment glint as it was conflicting with the hand's texture
+
 ## [v8.0.0-1.20.1] - 2023-06-27
 - Ported to Minecraft 1.20.1